Reffactor test_login module.
